+++ R/ggpie.R
@@ -7,8 +7,9 @@
 #' @param label_info Label information type, chosen from count, ratio and all (count and ratio). Default: count.
 #' @param label_type Label style, chosen from circle, horizon and none (no label). Default: circle.
 #' @param label_split Pattern used to split the label, support regular expression. Default: space.
+#' @param label_len The length of label text. Used when \code{label_split} is NULL. Default: 40.
 #' @param label_gap Gap between label and pie plot, used when \code{label_pos} is out.
-#' @param labal_threshold Threashold of the ratio to determine label position (in/out pie). Default: NULL.
+#' @param label_threshold Threshold of the ratio to determine label position (in/out pie). Default: NULL.
 #' @param label_pos Label position, chosen from in and out. Default: in.
 #' @param label_size Size of the label. Default: 4.
 #' @param label_color Color of the label. Default: black.
@@ -22,6 +23,7 @@
 #' @importFrom grDevices colorRampPalette
 #' @importFrom RColorBrewer brewer.pal
 #' @importFrom scales percent
+#' @importFrom stringr str_wrap
 #' @import ggplot2
 #' @importFrom ggrepel geom_text_repel
 #' @export
@@ -65,16 +67,16 @@
 #'   label_info = "all", label_type = "horizon", label_split = NULL,
 #'   label_size = 4, label_pos = "out"
 #' )
-#' # with label threashold
+#' # with label threshold
 #' ggpie(
 #'   data = diamonds, group_key = "cut", count_type = "full",
 #'   label_info = "all", label_type = "horizon", label_split = NULL,
-#'   label_size = 4, label_pos = "in", labal_threshold = 10
+#'   label_size = 4, label_pos = "in", label_threshold = 10
 #' )
 ggpie <- function(data, group_key = NULL, count_type = c("count", "full"), fill_color = NULL, label_info = c("count", "ratio", "all"),
-                  label_split = "[[:space:]]+", label_color = "black",
+                  label_split = "[[:space:]]+", label_len = 40, label_color = "black",
                   label_type = c("circle", "horizon", "none"), label_pos = c("in", "out"), label_gap = 0.05,
-                  labal_threshold = NULL, label_size = 4, border_color = "black", border_size = 1) {
+                  label_threshold = NULL, label_size = 4, border_color = "black", border_size = 1) {
   # check parameters
   count_type <- match.arg(arg = count_type)
   label_info <- match.arg(arg = label_info)
@@ -84,7 +86,7 @@ ggpie <- function(data, group_key = NULL, count_type = c("count", "full"), fill_
   # prepare plot data
   plot.data <- PrepareData(
     data = data, group_key = group_key, count_type = count_type, fill_color = fill_color,
-    label_info = label_info, label_split = label_split, label_color = label_color
+    label_info = label_info, label_split = label_split, label_len = label_len, label_color = label_color
   )
   data <- plot.data[["data"]]
   fill_color <- plot.data[["fill_color"]]
@@ -141,7 +143,7 @@ ggpie <- function(data, group_key = NULL, count_type = c("count", "full"), fill_
         scale_fill_manual(values = fill_color) +
         scale_colour_manual(values = label_color)
     } else if (label_pos == "in") {
-      if (is.null(labal_threshold)) {
+      if (is.null(label_threshold)) {
         pie_plot <- ggplot(data, aes(x = 1, y = Freq, fill = group)) +
           geom_bar(width = 1, stat = "identity", color = border_color, size = border_size) +
           geom_text_repel(
@@ -158,13 +160,13 @@ ggpie <- function(data, group_key = NULL, count_type = c("count", "full"), fill_
         pie_plot <- ggplot(data, aes(x = 1, y = Freq, fill = group)) +
           geom_bar(width = 1, stat = "identity", color = border_color, size = border_size) +
           geom_text_repel(
-            data = data[data$Freq < labal_threshold, ],
+            data = data[data$Freq < label_threshold, ],
             aes(label = label, y = CumFreq, x = after_stat(1.5), colour = group), show.legend = FALSE,
             size = label_size, point.padding = NA, max.overlaps = Inf, nudge_x = 1, nudge_y = 1,
             segment.curvature = -0.2, segment.ncp = 10, segment.angle = 20
           ) +
           geom_text(
-            data = data[data$Freq >= labal_threshold, ],
+            data = data[data$Freq >= label_threshold, ],
             aes(x = 1, y = CumFreq, label = label, colour = group), show.legend = FALSE,
             size = label_size
           ) +