+++ NEWS.md
@@ -1,9 +1,19 @@
+# ggpie 0.2.4
+* Added `label_len` to adjust the length of labels (Fix #2).
+* Corrected some spelling mistakes
+
+-----------
+
 # ggpie 0.2.0
-* Add `ggrosepie` to create rose **pie** and **donut** plot, it can also **support single or two variables**.
+* Added `ggrosepie` to create rose **pie** and **donut** plot, it can also **support single or two variables**.
+
+-----------
 
 # ggpie 0.1.0
 
-* Add `ggpie` to create 2D pie plot with single group variable.
-* Add `ggdonut` to create 2D donut plot with single group variable.
-* Add `ggnestedpie` to create 2D nested pie plot with two group variables.
-* Add `ggpie3D` to create 3D pie plot with single group variable.
+* Added `ggpie` to create 2D pie plot with single group variable.
+* Added `ggdonut` to create 2D donut plot with single group variable.
+* Added `ggnestedpie` to create 2D nested pie plot with two group variables.
+* Added `ggpie3D` to create 3D pie plot with single group variable.
+
+-----------